Fixed Cargo FMT and Clippy Errors.

+++ src/emulation_core/riscv/instruction.rs
@@ -119,7 +119,8 @@ impl TryFrom<u32> for Instruction {
             })),
 
             // I-type instructions:
-            OPCODE_IMM | OPCODE_IMM_32 | OPCODE_JALR | OPCODE_LOAD | OPCODE_MISC_MEM | OPCODE_SYSTEM => Ok(Instruction::IType(IType {
+            OPCODE_IMM | OPCODE_IMM_32 | OPCODE_JALR | OPCODE_LOAD | OPCODE_MISC_MEM
+            | OPCODE_SYSTEM => Ok(Instruction::IType(IType {
                 imm: (value >> 20) as u16,
                 rs1: ((value >> 15) & 0x1f) as u8,
                 funct3: ((value >> 12) & 0x07) as u8,
@@ -149,14 +150,14 @@ impl TryFrom<u32> for Instruction {
 
             // U-type instruction:
             OPCODE_LUI | OPCODE_AUIPC => Ok(Instruction::UType(UType {
-                imm: (value >> 12) as u32,
+                imm: value >> 12,
                 rd: ((value >> 7) & 0x1f) as u8,
                 op: (value & 0x7f) as u8,
             })),
 
             // J-type instruction:
             OPCODE_JAL => Ok(Instruction::JType(JType {
-                imm: (value >> 12) as u32,
+                imm: value >> 12,
                 rd: ((value >> 7) & 0x1f) as u8,
                 op: (value & 0x7f) as u8,
             })),
